Extract symbols
